% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-biofilm-sequence.R
\name{analyze_biofilm_sequence,bsocial-method}
\alias{analyze_biofilm_sequence,bsocial-method}
\title{Analyze Biofilm Assembly Sequences}
\usage{
\S4method{analyze_biofilm_sequence}{bsocial}(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object after \code{analyze_social_behavior()} has been called.}
}
\value{
The modified \linkS4class{bsocial} object with biofilm paths and plot functions.
}
\description{
Builds a directed graph of possible consortium assembly paths based on
strain subset relationships. Uses igraph to find shortest paths between
simpler and more complex consortia.
}
