% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_biofilm_sequence}
\alias{analyze_biofilm_sequence}
\title{Analyze Biofilm Assembly Sequences}
\usage{
analyze_biofilm_sequence(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\description{
Analyze Biofilm Assembly Sequences
}
