% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_diversity}
\alias{analyze_diversity}
\title{Analyze Diversity Effect on Fitness}
\usage{
analyze_diversity(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\description{
Analyze Diversity Effect on Fitness
}
