% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_growth}
\alias{analyze_growth}
\title{Analyze Growth Metrics}
\usage{
analyze_growth(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\description{
Analyze Growth Metrics
}
