% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_social_behavior}
\alias{analyze_social_behavior}
\title{Analyze Social Behavior (Fitness Effects)}
\usage{
analyze_social_behavior(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\description{
Analyze Social Behavior (Fitness Effects)
}
