% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{analyze_stability}
\alias{analyze_stability}
\title{Analyze Consortium Stability}
\usage{
analyze_stability(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}
}
\description{
Analyze Consortium Stability
}
