% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{calculate_growth_params}
\alias{calculate_growth_params}
\title{Calculate Growth Parameters from Preprocessed Curves}
\usage{
calculate_growth_params(.Object, ...)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}

\item{...}{Additional arguments (method).}
}
\description{
Calculate Growth Parameters from Preprocessed Curves
}
