% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-processed-curves.R
\name{plot_processed_curves,bsocial-method}
\alias{plot_processed_curves,bsocial-method}
\title{Plot Preprocessed Growth Curves}
\usage{
\S4method{plot_processed_curves}{bsocial}(.Object)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object after \code{transform_raw_data()} has been called.}
}
\value{
A ggplot2 object.
}
\description{
Creates a faceted line plot of mean growth curves after preprocessing,
grouped by experimental condition.
}
