% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{summarize_social_behavior}
\alias{summarize_social_behavior}
\title{Classify Strains as Cooperators, Cheaters, or Neutrals}
\usage{
summarize_social_behavior(.Object, ...)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}

\item{...}{Additional arguments.}
}
\description{
Classify Strains as Cooperators, Cheaters, or Neutrals
}
