% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsocial-class.R
\name{transform_raw_data}
\alias{transform_raw_data}
\title{Preprocess Raw Plate Reader Data}
\usage{
transform_raw_data(.Object, ...)
}
\arguments{
\item{.Object}{A \linkS4class{bsocial} object.}

\item{...}{Additional arguments (groups, bg_type, bg_param).}
}
\description{
Preprocess Raw Plate Reader Data
}
